Input Basic and Viewport
-Added GameViewport class to manage viewport creation, and update.
-Added Resize function.
-Added Generic InputManager class.
-Added BasicInput class to handle map movement. Zoom and pan.
-Cleaned some code.

+++ core/src/ipca/hrem/com/BasicResources/Grid.java
@@ -44,7 +44,7 @@ public class Grid {
     public void generateMap(){          //Generates firstMap
         for (int i = 0; i < gridWidth; i++) {
             for (int j = 0; j < gridHeight; j++) {
-                cells[i][j] = new GridCell(GridCell.CellType.test, new Vector2(i,j), (int)Map.CELL_SCALE);
+                cells[i][j] = new GridCell(GridCell.CellType.test, new Vector2(i,j), Map.CELL_SCALE);
             }
         }
     }
